mgw: Tear down all RTP flows to avoid race condition on tear down
When stopping the test TC_two_crcx_and_unsolicited_rtp the unsolicited
RTP stream is not stopped. As a result it could happen that between
tearing down the other flows and stopping the test an unsolicited RTP
packet is sent to a closed socket.
The resulting ICMP destination unreachable packet translates to a
"Connection refused" error on the sending socket and fails the test.

Avoid this by making sure the unsolicited RTP sender is stopped before
stopping the test.

Change-Id: Ied839596589609e75fa487046a85db48991e4c73