Osmocom_CTRL_Adapter: Let the OS decide over the local port number
At the moment the function f_ipa_ctrl_start() is starting the IPA
emulation client with parameter -1 for local port. This is internally
translated to port number 9999, which is a fixed number. This makes
it impossible to have two control interfaces at the same time.

Lets use 0 as local port, so that the OS is selecting a free port
number automatically.

Change-Id: Ie6648f8f4c1e065c174868c35eb64ee034ace3ce
Related: OS#3645