bts: TC_pcu_{time,rts}_ind: Higher tolerances
When executing the test on our build slaves, we ran into the following:

> Number of TDMA Frames (1100) not matching (1073 .. 1093)
> Number of TDMA Frames (1096) not matching (1073 .. 1093)

So it seems the tolerance was a bit too tight.  If the timer runs
for a bit more than the requested amount of seconds (e.g. due to
high system load), the number of TDMA frames is likely a bit higher
and hence we need to permit a larger tolerance.

Change-Id: Ib00bb7e7f223fb0329d00e3835127b8c31299e27