introduce MNCC_Emulation mp_mncc_version
Default the MNCC version to the current osmo-sip-connector's master branch MNCC
version.

As soon as the new version (6) is merged, we can bump it here to make tests for
master work again.

For 'latest' builds, we can adjust osmo-ttcn3-hacks to use version 5, and also
see those tests still passing.

Change-Id: I3eb6e0132dc99ebe41a98cc5c329ed4864b770d6