sccp: Modify SCCP_tests to be run against sccp_demo_user in srv mode
It's easier to handle several tests if they are the ones connecting to
the program under tests.
Forthcoming SCCP_Tests_RAW require use of this mode too.

Take the chance to tweak other values which were wrong, in order to have
CR answered correctly with a CC later on.

Related: OS#4343
Change-Id: Ibf99b1546411b9a9b4375b82f1003d93d3c1645d