MGCP_Test: fix test vector in TC_amr_oa_bwe_rtp_conversion
The rtp payload test vector in TC_amr_oa_bwe_rtp_conversion is wrong, it
lacks the last byte which should be 0x00. Also the testvector is not
very well chossen since it after BWE conversion the actual payload does
not shrink (even if it looks like if it would because of the 0x00 byte
at the end). Lets pick a better payload from a real world trace that
actually shrinks by one byte when it is converted to BWE and use that
one.

Change-Id: Id4256049bbca49ad5c2eb0579128838ebae062f8